pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
packages = ["src"]

[project]
name = "electric_toolbox"
version = "0.1.0"
description = "Personal Blog"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepdiff>=8.1.1",
    "expression[pydantic]>=5.3.0",
    "hypothesis>=6.123.9",
    "jinja2>=3.1.4",
    "markdown>=3.7",
    "mypy>=1.13.0",
    "pygments>=2.18.0",
    "pymdown-extensions>=10.13",
    "pytailwindcss>=0.2.0",
    "pytest>=8.3.4",
    "python-frontmatter==1.1.0",
    "python-slugify>=8.0.4",
    "types-markdown>=3.7.0.20241204",
]

[dependency-groups]
dev = [
    "flask>=3.1.0",
    "livereload>=2.7.0",
]

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]




# Same as Black.
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "W", "S", "PLC", "PLE", "PLR", "RUF"]

[tool.ruff.lint.pydocstyle]
convention = "google"

# Exclude certain rules in test folder
[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = ["S101", "F403", "PLR2004"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
strict = true
warn_unreachable = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_untyped_decorators = true
ignore_missing_imports = false
follow_imports = "normal"
